+++ qme/main/database/models.py
@@ -8,15 +8,15 @@ with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 """
 
+from qme.exceptions import MissingImportError
 import json
-import sys
 
 
 try:
     from sqlalchemy import Column, DateTime, String, Text, func
     from sqlalchemy.ext.declarative import declarative_base
 except:
-    sys.exit("sqlalchemy is required for a non-filesystem database.")
+    raise MissingImportError("sqlalchemy", "is required for a non-filesystem database.")
 
 # Shared declarative base
 Base = declarative_base()