+++ qme/main/database/sqlite.py
@@ -9,10 +9,10 @@ with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 """
 
 from qme.main.database.relational import RelationalDatabase
+from qme.exceptions import InvalidDatabaseFilename
 
 import logging
 import os
-import sys
 
 bot = logging.getLogger("qme.main.database.sqlite")
 
@@ -34,10 +34,7 @@ class SqliteDatabase(RelationalDatabase):
         # Derive database path, use default of qme.db if not provided
         db_path = os.path.join(config_dir, database_file)
         if not db_path.endswith(".db"):
-            bot.error(
-                f"Invalid database file for sqlite, {database_file} does not end in .db"
-            )
-            sys.exit(1)
+            raise InvalidDatabaseFilename(database_file)
 
         self.config = config
         self.db = "sqlite:///%s" % db_path