+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip.
 
 ## [0.2.x](https://github.com/vsoch/qme/tree/master) (0.0.x)
+ - adding exceptions specific to QueueMe over system exit (0.0.13)
  - using standard Python logging for error/other messages (0.0.12)
  - adding message argument to run/re-run, print of host/port to start (0.0.11)
  - first release with basic dashboard and databases (0.0.1)